# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ViperConnect API
  version: 2.5.2
  description: |
    API compatível com WhatsApp Cloud (formato de payload), utilizando Baileys como transporte.
    - Endpoints espelham em grande parte a estrutura do Cloud API (ex.: `v15.0/{phone}/messages`).
    - Suporta envios de texto, mídia (imagem/áudio/documento/vídeo), contatos e mensagens interativas (botões/listas).
servers:
  - url: /
security:
  - ApiToken: []
paths:
...
  /v15.0/{phone}/register:
    post:
      summary: Registrar/atualizar configuração da sessão
      parameters:
        - $ref: '#/components/parameters/PhonePath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterConfig'
      responses:
        '200':
          description: Configuração aplicada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterConfig'
        '400':
          description: Erro ao aplicar configuração
...
components:
  securitySchemes:
    ApiToken:
      type: apiKey
      in: header
      name: Authorization
      description: 'API token. Ex.: 1 ou Bearer <token>'
  parameters:
    PhonePath:
      name: phone
      in: path
      required: true
      schema:
        type: string
      description: 'Número (ID) da sessão, ex.: 5511999999999'
  schemas:
...
    RegisterConfig:
      description: 'Configuração da sessão. Campos variam por provedor.'
      type: object
      additionalProperties: true
      properties:
        authToken:
          type: string
        provider:
          type: string
          enum: [baileys, zapo]
          default: baileys
          description: Motor WhatsApp da sessão. Ao alterar de Baileys para Zapo, a autenticação compatível é migrada automaticamente sem apagar a origem.
        connectionType:
          type: string
...